add improvement based on comments
